+++ embedded-nal-async/src/dns.rs
@@ -1,6 +1,5 @@
 use crate::IpAddr;
 use embedded_nal::AddrType;
-use heapless::String;
 
 /// This trait is an extension trait for [`TcpStack`] and [`UdpStack`] for dns
 /// resolutions. It does not handle every DNS record type, but is meant as an
@@ -24,13 +23,22 @@ pub trait Dns {
 		addr_type: AddrType,
 	) -> Result<IpAddr, Self::Error>;
 
-	/// Resolve the hostname of a host, given its ip address
+	/// Resolve the hostname of a host, given its ip address.
+	///
+	/// The hostname is stored at the beginning of `result`, the length is returned.
+	///
+	/// If the buffer is too small to hold the domain name, an error should be returned.
 	///
 	/// **Note**: A fully qualified domain name (FQDN), has a maximum length of
-	/// 255 bytes [`rfc1035`]
+	/// 255 bytes according to [`rfc1035`]. Therefore, you can pass a 255-byte long
+	/// buffer to guarantee it'll always be large enough.
 	///
 	/// [`rfc1035`]: https://tools.ietf.org/html/rfc1035
-	async fn get_host_by_address(&self, addr: IpAddr) -> Result<String<256>, Self::Error>;
+	async fn get_host_by_address(
+		&self,
+		addr: IpAddr,
+		result: &mut [u8],
+	) -> Result<usize, Self::Error>;
 }
 
 impl<T: Dns> Dns for &T {
@@ -44,7 +52,11 @@ impl<T: Dns> Dns for &T {
 		T::get_host_by_name(self, host, addr_type).await
 	}
 
-	async fn get_host_by_address(&self, addr: IpAddr) -> Result<String<256>, Self::Error> {
-		T::get_host_by_address(self, addr).await
+	async fn get_host_by_address(
+		&self,
+		addr: IpAddr,
+		result: &mut [u8],
+	) -> Result<usize, Self::Error> {
+		T::get_host_by_address(self, addr, result).await
 	}
 }