+++ src/lib.rs
@@ -8,9 +8,6 @@ mod dns;
 mod stack;
 
 pub use nb;
-// Needed by embedded-nal trait implementers who build get_host_by_address results, or by trait
-// users who pass the results on.
-pub use heapless;
 
 #[cfg(not(any(feature = "ip_in_core", feature = "no-std-net")))]
 compile_error!("You must select the ip_in_core feature or the no-std-net feature");